Create and exchange account-global keypairs
Things like display names and status messages aren't directed at any
single user, so they can't be encrypted with an account-pair specific
key. Instead, create a keypair as part of WHAccount which is used for
these, and pass this public key along with the account-pair specific
public key when doing key exchange.