Fix bug that results in contacts sometimes not being created on one device
Disconnecting immediately after receiving the last bit of data resulted
in the other peer sometimes receiving the disconnect notification before
reading the sent data and cancelling the key exchange immediately before
it completed. Fix this by sending a dummy packet at the endand only
disconnecting when one of the peers receives that (and not cancelling if
the session is disconnected instead).